+++ src/components/screens/game-screens/InventoryScreen.vue
@@ -20,7 +20,7 @@
 </template>
 
 <script lang="ts">
-	import useMonsterSlayerService from "@/services/MonsterSlayerFactory.vue";
+	import useMonsterSlayerService from "@/services/monster-slayer-service";
     import { IItem } from "@/store/types";
     import { computed, defineComponent, ref } from "vue";
