+++ src/components/screens/game-screens/SkillsScreen.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script lang="ts">
-	import useMonsterSlayerService from "@/services/MonsterSlayerFactory.vue";
+	import useMonsterSlayerService from "@/services/monster-slayer-service";
     import { ISkills } from "@/store/types";
     import { defineComponent, ref } from "vue";
