--- src/services/service-container.ts
@@ -1,5 +1,0 @@
-// /* eslint-disable @typescript-eslint/no-empty-function */
-// import MonsterSlayerService from '../services/monster-slayer.factory';
-// const serviceInstance = new MonsterSlayerService();
-
-// export default serviceInstance;
